requirements.txt
google-generativeai==0.4.0
pillow==10.2.0
streamlit==1.31.1
streamlit-option-menu==0.3.12
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.13